// src/modules/doctor/domain/enums/SpecialtyEnum.ts
export enum SpecialtyEnum {
  GENERAL_MEDICINE = 'general_medicine',
  CARDIOLOGY = 'cardiology',
  DERMATOLOGY = 'dermatology',
  ORTHOPEDICS = 'orthopedics',
  PEDIATRICS = 'pediatrics',
  GYNECOLOGY = 'gynecology',
  NEUROLOGY = 'neurology',
  OPHTHALMOLOGY = 'ophthalmology',
  ENT = 'ent',
  PSYCHIATRY = 'psychiatry',
  DENTISTRY = 'dentistry',
  UROLOGY = 'urology',
  GASTROENTEROLOGY = 'gastroenterology',
  PULMONOLOGY = 'pulmonology',
  ENDOCRINOLOGY = 'endocrinology',
}